Move Disposable and DisposeBag their own file Disposable.swift
